+++ release/src/router/inadyn/plugins/dyndns.c
@@ -81,6 +81,21 @@ static ddns_system_t dynsip = {
 	.server_url   = "/nic/update"
 };
 
+/* Support HOSTNAME.selfhost.eu, see issue #215 */
+static ddns_system_t selfhost = {
+	.name         = "default@selfhost.de",
+
+	.request      = (req_fn_t)request,
+	.response     = (rsp_fn_t)response,
+
+	.checkip_name = DYNDNS_MY_IP_SERVER,
+	.checkip_url  = DYNDNS_MY_CHECKIP_URL,
+	.checkip_ssl  = DYNDNS_MY_IP_SSL,
+
+	.server_name  = "carol.selfhost.de",
+	.server_url   = "/nic/update"
+};
+
 static ddns_system_t noip = {
 	.name         = "default@no-ip.com",
 
@@ -231,6 +246,7 @@ PLUGIN_INIT(plugin_init)
 	plugin_register(&dyndns);
 	plugin_register(&dnsomatic);
 	plugin_register(&dynsip);
+	plugin_register(&selfhost);
 	plugin_register(&noip);
 	plugin_register(&_3322);
 	plugin_register(&henet);
@@ -247,6 +263,7 @@ PLUGIN_EXIT(plugin_exit)
 	plugin_unregister(&dyndns);
 	plugin_unregister(&dnsomatic);
 	plugin_unregister(&dynsip);
+	plugin_unregister(&selfhost);
 	plugin_unregister(&noip);
 	plugin_unregister(&_3322);
 	plugin_unregister(&henet);